Refactor log content to separate component for reuse.
